==> codigo/Scripts/1.0_Procesamiento_y_Analisis_de_Encuesta.R <==
# PAQUETES Y DATOS -------------------------------------------------------------
library(RColorBrewer)
library(sf)
library(tidyverse)
Encuesta <- read_csv("datos/Encuesta.csv")


# LIMPIEZA DE DATOS ------------------------------------------------------------

# Si se quiere comprobar las clases de cada columna
lapply(Encuesta,class)
str(Encuesta)


# Cambiar la clase de "Edad" y "Frecuencia" (caracter a numérico)
Encuesta$Edad <- as.numeric(Encuesta$Edad)
Encuesta$Frecuencia <- as.numeric(Encuesta$Frecuencia)

#Cambiar "Formacion", Interes y Uso plat. a factor
Encuesta$Formacion <- as.factor(Encuesta$Formación)
Encuesta$`Uso de plataforma` <- as.factor(Encuesta$`Uso de plataforma`)
Encuesta$Interes_Naturaleza <- as.factor(Encuesta$Interes_Naturaleza)

# Cambiar las columnas de taxones a factor
Encuesta$Aves <- as.factor(Encuesta$Aves)
Encuesta$Anfibios <- as.factor(Encuesta$Anfibios)
Encuesta$Reptiles <- as.factor(Encuesta$Reptiles)
Encuesta$Mamíferos <- as.factor(Encuesta$Mamíferos)
Encuesta$Arácnidos <- as.factor(Encuesta$Arácnidos)
Encuesta$Insectos <- as.factor(Encuesta$Insectos)
Encuesta$Plantas <- as.factor(Encuesta$Plantas)
Encuesta$Hongos <- as.factor(Encuesta$Hongos)


# Eliminar los NA
Encuesta <- na.omit(Encuesta)


# Estadisticas
summary(Encuesta)



saveRDS(Encuesta, "datos/Encuesta.rds")


# GRAFICOS ---------------------------------------------------------------------

# Histograma de edades
ggplot(Encuesta) +
  geom_histogram(aes(x=Edad), bins = 10, fill= "skyblue", col= "white") + 
  theme_gray()


# Afinidad por la naturaleza en cada departamento
ggplot(Encuesta) +
  geom_bar(mapping = aes(x=Afinidad_Naturaleza, fill=Departamento),  
           show.legend = F,) + 
  scale_fill_brewer(palette = "Spectral") +
  labs(title = "Afinidad por la naturaleza",
       x = "Afinidad", y = "Participantes") + 
  facet_grid(~Departamento) +
  theme_gray()

   ## Afinidad por la naturaleza en función de la edad
   ggplot(Encuesta) + 
     geom_boxplot(aes(x=Afinidad_Naturaleza, y=Edad, fill=Departamento),
                  show.legend = F) +
     scale_fill_brewer(palette = "Spectral") +
     facet_wrap(~Departamento) +
     xlim(1,5) +
     labs(y='Edad', x='Afinidad por la Naturaleza') +
     theme_gray()


# Afinidad por la tecnologia en cada departamento
ggplot(Encuesta) +
  geom_bar(mapping = aes(x=Afinidad_Tecnologia, fill=Departamento),
           show.legend = F,) + 
  labs(title = "Afinidad por la Tecnologia",
       x = "Afinidad", y = "Participantes") + 
  scale_fill_brewer(palette = "Spectral") + 
  facet_grid(~Departamento) +
  theme_gray()

   ## Afinidad por la tecnologia en función de la edad
   ggplot(Encuesta) + 
     geom_boxplot(aes(x=Afinidad_Tecnologia, y=Edad, fill=Departamento),
                  show.legend = F) +
     scale_fill_brewer(palette = "Spectral") +
     facet_wrap(~Departamento) +
     labs(y='Edad', x='Afinidad por la Tecnologia') +
     theme_gray()


# Frecuencia de salida a la Naturaleza 
ggplot(Encuesta) +
  geom_bar(mapping = aes(x=Frecuencia, fill=Departamento), show.legend = F) + 
  labs(title = "Frecuencia de salida a la Naturaleza",
       x = "Frecuencia", y = "Participantes") + 
  scale_fill_brewer(palette = "Spectral") + 
  facet_wrap(~Departamento) +
  theme_gray()

   ## Frecuencia de salidas en funcion de la edad
   ggplot(Encuesta) + 
     geom_boxplot(aes(x=Frecuencia, y=Edad, fill=Departamento),
                  show.legend = F) +
     scale_fill_brewer(palette = "Greens") +
     facet_wrap(~Departamento) +
     labs(y='Edad', x='Frecuencia de salidas') +
     theme_gray()
